ASoC: wcd-mbhc: Fix incorrect detection during slow plug insertion
When any plug is inserted into our jack very slowly, it can be detected as
a plug with the ground and mic pins swapped, causing MBHC to give up early
and skew the plug detection. Instead of giving up when this happens, keep
iterating to make sure the plug is properly detected.

Signed-off-by: Sultanxda <sultanxda@gmail.com>