ASoC: wcd-mbhc: Ignore button interrupts until detection is complete
When a plug is inserted into the jack quickly, spurious button
interrupts are generated and interfere with UX for a short amount of
time after plug detection is complete. To reproduce: play music out of
speaker, plug in jack during music playback, and the music playback will
pause after being routed to the peripheral inserted into the jack (due
to a spurious button press).

To fix this, ignore all button interrupts during plug detection, as well
as for 750 ms after plug detection is completed.

Signed-off-by: Sultanxda <sultanxda@gmail.com>